[Feat] 일정 시간동안 유지되는 아이템 구현, 우측 상단에 효과 지속 시간UI 표시

+++ Assets/Scripts/Base/BaseCharacter.cs
@@ -7,18 +7,18 @@ public abstract class BaseCharacter : MonoBehaviour
     [Header("스탯 설정")]
     [SerializeField] protected float maxHp = 0;
     [SerializeField] protected float hpPassive = 5;
-    [SerializeField] protected int atkPower = 0;
-    [SerializeField] protected int defPower = 0;
-    [SerializeField] protected int walkSpeed = 0;
-    [SerializeField] protected int runSpeed = 0;
+    [SerializeField] protected float atkPower = 0;
+    [SerializeField] protected float defPower = 0;
+    [SerializeField] protected float walkSpeed = 0;
+    [SerializeField] protected float runSpeed = 0;
     protected float curHp = 0;
     public float CurHp { get { return curHp; } set { curHp = Mathf.Clamp(value, 0f, maxHp); } }
     public float MaxHp { get { return maxHp; } set { maxHp = value; } }
     public float HpPassive { get { return hpPassive; } }
-    public int AtkPower { get { return atkPower; } set { atkPower = value; } }
-    public int DefPower { get { return defPower; } set { defPower = value; } }
-    public int WalkSpeed { get { return walkSpeed; } set { walkSpeed = value; } }
-    public int RunSpeed { get { return runSpeed; } set { runSpeed = value; } }
+    public float AtkPower { get { return atkPower; } set { atkPower = value; } }
+    public float DefPower { get { return defPower; } set { defPower = value; } }
+    public float WalkSpeed { get { return walkSpeed; } set { walkSpeed = value; } }
+    public float RunSpeed { get { return runSpeed; } set { runSpeed = value; } }
     
 
     // 컨트롤러도 여기에?
